requirements.txt
streamlit
yt-dlp
google-generativeai